Fixed remove devices throughout from device list
For now we mark devices not in the device list as ’removed by the
server’. This is a bit more complicated as other functions need to know
that these device aren’t really on the server anymore. But it seems
wrong that the server can force you to delete information.  #579