Remove threaded operation
The caching code (in the else branch of the "if (reset)") would do some
calculations on a worker thread (but not the actual size calculation,
that was commented out) and only after that call the doReload() block.
This could lead to a race condition where "didReceiveChanges" was
called before the "doReload" was called, resulting in a crash.